Move emitter to lib/methods/helpers, closer to events.ts
